requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==11.0.0
djangorestframework_simplejwt==5.4.0
celery==5.4.0
requests==2.32.3
django-celery-beat==2.7.0
eventlet==0.39.0
django-cors-headers==4.7.0
drf-yasg==1.21.8
redis==5.2.1